feat: Organize structure
- Introduced ESLint configuration to enforce coding standards and best practices for JavaScript files.
- Added jsconfig.json to specify compiler options for ES2022 and enable type checking for .js files.
- Implemented a new VS Code extension for generating responsive images, including commands for image processing and completion providers for responsive image tags.
- Added functionality to prompt users for image files, output directories, and sizes to generate, with error handling and user feedback.
- Integrated sharp library for image resizing and processing, supporting multiple output sizes and relative path configurations.
