+++ package.json
@@ -25,14 +25,36 @@
     "onLanguage:csharp",
     "onLanguage:cshtml"
   ],
-  "main": "./extension.js",
+  "main": "./src/extension.js",
   "contributes": {
     "commands": [
       {
         "command": "responsive-image-generator.generate",
         "title": "Generate Responsive Image"
       }
-    ]
+    ],
+    "configuration": {
+      "type": "object",
+      "title": "Responsive Image Generator",
+      "properties": {
+        "responsiveImageGenerator.defaultSizes": {
+          "type": "array",
+          "items": { "type": "number" },
+          "default": [320, 480, 640, 960, 1280, 1600, 1920, 2560, 3840, 5120, 7680],
+          "description": "Default image sizes to generate."
+        },
+        "responsiveImageGenerator.useRelativePaths": {
+          "type": "boolean",
+          "default": true,
+          "description": "Insert image paths as relative to ```staticAssetsRoot``` root."
+        },
+        "responsiveImageGenerator.staticAssetsRoot": {
+          "type": "string",
+          "default": "",
+          "description": "Default working directory for your project. If you're using relative paths, this is the folder relative to which paths will be generated."
+        }
+      }
+    }
   },
   "scripts": {
     "lint": "eslint .",
